feat: refactor page name, fix bug
